Struct rstox::core::Tox

pub struct Tox {
    // some fields omitted
}

Methods

impl Tox

fn new(opts: ToxOptions, data: Option<&[u8]>) -> Result<Tox, InitError>

Create a new tox instance

fn iter(&mut self) -> ToxIter

Ticks the Tox and returns an iterator to the Tox events

fn tick(&mut self)

This function animates tox by calling tox_do() It function should be called at least several times per second. Use wait() method to get optimal delays

fn wait(&self)

This function makes thread sleep for a some time, optimal for tick() method

fn bootstrap(&mut self, host: &str, port: u16, public_key: PublicKey) -> Result<(), BootstrapError>

Sends a "get nodes" request to the given bootstrap node with IP, port, and public key to setup connections.

This function will attempt to connect to the node using UDP and TCP at the same time.

Tox will use the node as a TCP relay in case Tox_Options.udp_enabled was false, and also to connect to friends that are in TCP-only mode. Tox will also use the TCP connection when NAT hole punching is slow, and later switch to UDP if hole punching succeeds.

Panics

Panics if host string contains \0.

fn get_connection_status(&self) -> Connection

Get self connection status

fn get_address(&self) -> Address

Get self tox address

fn get_nospam(&self) -> [u8; 4]

Get self nospam

fn set_nospam(&mut self, nospam: [u8; 4])

Set self nospam

fn get_public_key(&self) -> PublicKey

Get self public key

fn set_name(&mut self, name: &str) -> Result<(), SetInfoError>

Set the nickname for the Tox client

fn get_name(&self) -> String

Get self nickname

fn set_status_message(&mut self, message: &str) -> Result<(), SetInfoError>

Set self status message

fn get_status_message(&self) -> String

Get self status message

fn set_status(&mut self, status: UserStatus)

Set self status

fn get_status(&self) -> UserStatus

Get self status

fn add_friend(&mut self, address: &Address, message: &str) -> Result<(), FriendAddError>

Add a friend to the friend list and send a friend request.

A friend request message must be at least 1 byte long and at most TOX_MAX_FRIEND_REQUEST_LENGTH.

Friend numbers are unique identifiers used in all functions that operate on friends. Once added, a friend number is stable for the lifetime of the Tox object. After saving the state and reloading it, the friend numbers may not be the same as before. Deleting a friend creates a gap in the friend number set, which is filled by the next adding of a friend. Any pattern in friend numbers should not be relied on.

fn add_friend_norequest(&mut self, address: &PublicKey) -> Result<(), FriendAddError>

Add a friend without sending a friend request.

This function is used to add a friend in response to a friend request. If the client receives a friend request, it can be reasonably sure that the other client added this client as a friend, eliminating the need for a friend request.

This function is also useful in a situation where both instances are controlled by the same entity, so that this entity can perform the mutual friend adding. In this case, there is no need for a friend request, either.

fn delete_friend(&mut self, fnum: u32) -> Result<(), ()>

Remove a friend from the friend list.

This does not notify the friend of their deletion. After calling this function, this client will appear offline to the friend and no communication can occur between the two.

fn friend_by_public_key(&self, public_key: PublicKey) -> Option<u32>

fn friend_exists(&self, fnum: u32) -> bool

fn get_friend_list(&self) -> Vec<u32>

fn get_friend_public_key(&self, fnum: u32) -> Option<PublicKey>

fn send_friend_message(&mut self, fnum: u32, kind: MessageType, message: &str) -> Result<u32, FriendSendMessageError>

Send a text chat message to an online friend.

This function creates a chat message packet and pushes it into the send queue.

The message length may not exceed TOX_MAX_MESSAGE_LENGTH. Larger messages must be split by the client and sent as separate messages. Other clients can then reassemble the fragments. Messages may not be empty.

The return value of this function is the message ID. If a read receipt is received, the triggered friend_read_receipt event will be passed this message ID.

Message IDs are unique per friend. The first message ID is 0. Message IDs are incremented by 1 each time a message is sent. If UINT32_MAX messages were sent, the next message ID is 0.

fn add_groupchat(&mut self) -> Result<i32, ()>

fn del_groupchat(&mut self, groupnumber: i32) -> Result<(), ()>

fn group_peername(&self, groupnumber: i32, peernumber: i32) -> Option<String>

fn group_peer_pubkey(&self, groupnumber: i32, peernumber: i32) -> Option<PublicKey>

fn invite_friend(&mut self, friendnumber: i32, groupnumber: i32) -> Result<(), ()>

fn join_groupchat(&mut self, friendnumber: i32, data: &[u8]) -> Result<i32, ()>

fn group_message_send(&mut self, groupnumber: i32, msg: &str) -> Result<(), ()>

fn group_action_send(&mut self, groupnumber: i32, act: &str) -> Result<(), ()>

fn group_number_peers(&self, groupnumber: i32) -> Option<i32>

fn group_get_names(&self, groupnumber: i32) -> Result<Vec<Option<String>>, ()>

fn group_get_title(&self, groupnumber: i32) -> Option<String>

fn group_set_title(&mut self, groupnumber: i32, title: &str) -> Result<(), ()>

fn count_chatlist(&self) -> u32

fn get_chatlist(&self) -> Vec<i32>

fn group_get_type(&self, groupnumber: i32) -> Option<GroupchatType>

fn save(&self) -> Vec<u8>

Get all all information associated with the tox instance as a Vec<u8>